pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp2mqtt"
version = "0.1.0"
description = "MCP MQTT Service for PWM Control and Device Communication"
readme = "README_PYPI.md"  # 使用PyPI专用README
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "mcp2everything", email = "mcp2everything@gmail.com" },
]
maintainers = [
    { name = "mcp2everything", email = "mcp2everything@gmail.com" },
]
keywords = [
    "mcp",
    "mqtt",
    "pwm",
    "control",
    "device",
    "communication",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyyaml>=6.0.1",
    "paho-mqtt>=1.6.1",
    "mcp-python>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/mcp2everything/mcp2mqtt"
Repository = "https://github.com/mcp2everything/mcp2mqtt.git"
Issues = "https://github.com/mcp2everything/mcp2mqtt/issues"
Changelog = "https://github.com/mcp2everything/mcp2mqtt/blob/main/CHANGELOG.md"

[project.scripts]
mcp2mqtt = "mcp2mqtt:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
]
